test(canon): align optional handler regressions
